refactor(inspector): extract dataflow panel
